+++ tests/test_slots.py
@@ -16,7 +16,7 @@ def test_slots() -> None:
     assert _check_slots(pc.Seq(()))
     assert _check_slots(pc.Set(()))
     assert _check_slots(pc.SetMut(()))
-    assert _check_slots(pc.Dict[str, str].new())
+    assert _check_slots(pc.Dict[str, str](()))
     assert _check_slots(pc.Some(42))
     assert _check_slots(pc.Range(0, 1))
     assert _check_slots(pc.Vec(()))